fix: Delete output directory correctly when table name is used
AB#55494
